Replicating from Playerstate launch now happens on clients (Launch needs to be a server function) Health HUD Modified
Playerstate is replicating damage and lives. HUD function for health has been modified to access the playerstate for its value because thats proper. And Added a multicast function to healthcomponent to launch players over the serverPlayerstate is replicating damage and lives. HUD function for health has been modified to access the playerstate for its value because thats proper. And Added a multicast function to healthcomponent to launch players over the server